+++ server/dataStructure/skiplist/SkipListNode.cpp
@@ -4,7 +4,7 @@
 
 #include "SkipListNode.h"
 
-SkipListNode::SkipListNode(const SkipListType& type, void* obj, double score, unsigned int level) : type(type) {
+SkipListNode::SkipListNode(const SkipListType& type, void* obj, double score, uint8_t level) : type(type) {
     this->obj = obj;
     this->score = score;
     for (int i = 0; i < level; i++) {
